backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Multi-Agent System with Native LLM Support
crewai>=0.28.8

# LLM APIs (Optional - CrewAI handles providers natively)
# openai>=1.12.0
# anthropic>=0.18.0
# ollama>=0.1.7

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.7  # PostgreSQL
aiosqlite>=0.19.0  # SQLite async support

# Cache
redis>=5.0.1
aioredis>=2.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Utilities
loguru>=0.7.2
httpx>=0.25.2
aiofiles>=23.2.1
python-dateutil>=2.8.2
pyyaml>=6.0.1

# Chinese Text Processing
jieba>=0.42.1

# Data Science & ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Development Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
pylint>=3.0.3
mypy>=1.7.0

# Monitoring
prometheus-client>=0.19.0